+++ src/include/iceberg/io/local_file_io.hh
@@ -1,11 +1,53 @@
 #pragma once
 
-#include "iceberg/io/filesystem.hh"
+#include "iceberg/io/file_io.hh"
 #include "iceberg/util/macros.hh"
 #include "iceberg/util/visibility.hh"
 
 namespace iceberg {
 namespace io {
+
+class ICEBERG_EXPORT SeekableFileInputStream : public SeekableInputStream {
+ public:
+  explicit SeekableFileInputStream(int fd) : fd_(fd){};
+  ~SeekableFileInputStream();
+
+  Status Close() override;
+
+  Result<int64_t> Tell() const override;
+
+  Result<int64_t> Read(int64_t nbytes, void* out) override;
+
+  Status Seek(int64_t position) override;
+
+  bool closed() const override;
+
+ private:
+  int fd_ = -1;
+  Status CheckClosed() const;
+  ICEBERG_DISALLOW_COPY_AND_ASSIGN(SeekableFileInputStream);
+};
+
+class ICEBERG_EXPORT PositionFileOutputStream : public PositionOutputStream {
+ public:
+  explicit PositionFileOutputStream(int fd) : fd_(fd){};
+  ~PositionFileOutputStream();
+
+  Status Close() override;
+
+  Result<int64_t> Tell() const override;
+
+  Status Write(const void* data, int64_t nbytes) override;
+
+  Status Flush() override;
+
+  bool closed() const override;
+
+ private:
+  int fd_ = -1;
+  Status CheckClosed() const;
+  ICEBERG_DISALLOW_COPY_AND_ASSIGN(PositionFileOutputStream);
+};
 
 /// \brief An local implementation of InputFile.
 class ICEBERG_EXPORT LocalInputFile : public InputFile {
@@ -39,15 +81,15 @@ class ICEBERG_EXPORT LocalOutputFile : public OutputFile {
   ICEBERG_DISALLOW_COPY_AND_ASSIGN(LocalOutputFile);
 };
 
-class ICEBERG_EXPORT LocalFileSystem : public FileSystem {
+class ICEBERG_EXPORT LocalFileIO : public FileIO {
  public:
-  LocalFileSystem() = default;
+  LocalFileIO() = default;
 
-  ~LocalFileSystem();
+  ~LocalFileIO();
 
-  std::string fs_name() const override { return "local"; }
+  std::string name() const override { return "local"; }
 
-  bool Equals(const FileSystem& other) const override;
+  bool Equals(const FileIO& other) const override;
 
   Result<std::shared_ptr<InputFile>> newInputFile(const std::string& path) override;
 
@@ -56,7 +98,7 @@ class ICEBERG_EXPORT LocalFileSystem : public FileSystem {
   Status DeleteFile(const std::string& path) override;
 
  private:
-  ICEBERG_DISALLOW_COPY_AND_ASSIGN(LocalFileSystem);
+  ICEBERG_DISALLOW_COPY_AND_ASSIGN(LocalFileIO);
 };
 
 }  // namespace io
