--- src/io/filesystem.cc
@@ -1,14 +1,0 @@
-#include "iceberg/io/filesystem.hh"
-
-namespace iceberg {
-namespace io {
-
-Status InputFile::CheckExists() const {
-  if (!exists()) {
-    return Status::Invalid("Input file not exists");
-  }
-  return Status::OK();
-}
-
-}  // namespace io
-}  // namespace iceberg
