+++ src/io/local_file_io.cc
@@ -1,11 +1,14 @@
-#include "iceberg/io/file.hh"
+#include "iceberg/io/local_file_io.hh"
 
 #include <fcntl.h>
-#include <stdlib.h>
 #include <sys/stat.h>
 #include <unistd.h>
 #include <cmath>
+#include <filesystem>
+#include <memory>
+#include <system_error>
 
+#include "iceberg/result.hh"
 #include "iceberg/util/logging.hh"
 #include "iceberg/util/macros.hh"
 
@@ -144,6 +147,84 @@ bool PositionFileOutputStream::closed() const { return fd_ == -1; }
 Status PositionFileOutputStream::CheckClosed() const {
   if (closed()) {
     return Status::Invalid("Invalid operation on closed file");
+  }
+  return Status::OK();
+}
+
+Result<int64_t> LocalInputFile::getLength() {
+  if (!exists()) {
+    return Status::Invalid("File not exists.");
+  }
+  return std::filesystem::file_size(location_);
+}
+
+Result<std::shared_ptr<SeekableInputStream>> LocalInputFile::newStream() {
+  ICEBERG_RETURN_NOT_OK(CheckExists());
+  int fd = open(location().c_str(), O_RDONLY);
+  if (fd < 0) {
+    return Status::IOError("Failed to open local file '", location(), "' errno: ", errno);
+  }
+  struct stat st;
+  int ret = fstat(fd, &st);
+  if (ret == 0 && S_ISDIR(st.st_mode)) {
+    return Status::IOError("Cannot open for reading: path '", location(),
+                           "' is a directory");
+  }
+  return std::make_shared<SeekableFileInputStream>(fd);
+}
+
+bool LocalInputFile::exists() const { return std::filesystem::exists(location()); }
+
+Result<std::shared_ptr<PositionOutputStream>> LocalOutputFile::create() {
+  if (std::filesystem::exists(location())) {
+    return Status::AlreadyExists("output file ", location(), " already exisits");
+  }
+
+  int fd = open(location().c_str(), O_CREAT | O_WRONLY, 0666);
+  if (fd == -1) {
+    return Status::IOError("Failed to open local file '", location(),
+                           "', errno: ", errno);
+  }
+  return std::make_shared<PositionFileOutputStream>(fd);
+}
+
+Result<std::shared_ptr<PositionOutputStream>> LocalOutputFile::createOrOverwrite() {
+  int fd = open(location().c_str(), O_CREAT | O_WRONLY | O_TRUNC, 0666);
+  if (fd == -1) {
+    return Status::IOError("Failed to open local file '", location(),
+                           "', errno: ", errno);
+  }
+  return std::make_shared<PositionFileOutputStream>(fd);
+}
+
+Result<std::shared_ptr<InputFile>> LocalOutputFile::toInputFile() const {
+  return std::make_shared<LocalInputFile>(location());
+}
+
+bool LocalFileIO::Equals(const FileIO& other) const {
+  if (other.name() != name()) {
+    return false;
+  }
+
+  return true;
+}
+
+LocalFileIO::~LocalFileIO() {}
+
+Result<std::shared_ptr<InputFile>> LocalFileIO::newInputFile(const std::string& path) {
+  return std::make_shared<LocalInputFile>(path);
+}
+
+Result<std::shared_ptr<OutputFile>> LocalFileIO::newOutputFile(const std::string& path) {
+  return std::make_shared<LocalOutputFile>(path);
+}
+
+Status LocalFileIO::DeleteFile(const std::string& path) {
+  std::error_code ec;
+  bool ret = std::filesystem::remove(path, ec);
+  if (!ret) {
+    return Status::IOError("Delete file '", path,
+                           "' failed, error message: ", ec.message());
   }
   return Status::OK();
 }
